% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEpathways-package.R
\docType{package}
\name{iSEEpathways-pkg}
\alias{iSEEpathways}
\alias{iSEEpathways-package}
\title{iSEE extension for panels related to pathway analysis}
\description{
\code{iSEEpathways} is a package that provides panels for \pkg{iSEE},
facilitating the interactive visualisation of pathway analysis results.
}
\examples{
library("iSEEde")
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/iSEE/iSEEpathways}
  \item Report bugs at \url{https://support.bioconductor.org/t/iSEEpathways}
}

}
\author{
Kevin Rue-Albrecht \email{kevin.rue-albrecht@imm.ox.ac.uk}
}
\keyword{internal}
