% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfgseaResults-class.R
\docType{methods}
\name{iSEEpathwaysResults-class}
\alias{iSEEpathwaysResults-class}
\alias{iSEEpathwaysResults}
\alias{show,iSEEpathwaysResults-method}
\title{The iSEEpathwaysResults class}
\description{
The iSEEpathwaysResults is a virtual class for storing a set of pathway analysis results.
}
\section{Slot overview}{

This class inherits all its slots directly from its parent class \code{DataFrame}.
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \code{\linkS4class{iSEEpathwaysResults}} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{pathwayType}(x)} returns the type of pathways stored in \code{x}.
}
}

\examples{
showClass("iSEEpathwaysResults")
}
\author{
Kevin Rue-Albrecht
}
