% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{pathway-generics}
\alias{pathway-generics}
\alias{featuresStats}
\alias{pathwayType}
\alias{pathwaysList}
\title{Generics for Pathway Analysis Results}
\description{
An overview of the generics for accessing common pieces of information in pathway analysis results.
}
\section{Definitions}{

\itemize{
\item \code{pathwayType(x)} returns a character scalar indicating the type of pathways analysed.
\item \code{pathwaysList(x)} returns the named list of pathways used in the analysis, or \code{NULL}.
\item \code{featuresStats(x)} returns the named numeric vector of feature-level statistics used in the analysis, or \code{NULL}.
}
}

\examples{
showMethods(pathwayType)
showMethods(pathwaysList)
showMethods(featuresStats)
}
\author{
Kevin Rue-Albrecht
}
