% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-SummarizedExperiment.R
\name{pathwaysResultsNames}
\alias{pathwaysResultsNames}
\alias{pathwaysResults}
\title{Extract contrast results embedded in a SummarizedExperiment object}
\usage{
pathwaysResultsNames(object)

pathwaysResults(object, name)
}
\arguments{
\item{object}{A \linkS4class{SummarizedExperiment} object.}

\item{name}{(Optional) Name of a single pathway analysis result name to extract.
Use \code{pathwaysResultsNames(object)} to list available names.}
}
\value{
For \code{pathwaysResultsNames}: the names of embedded contrast results available.

For \code{pathwaysResults}: a \code{list} of differential expression statistics.

If \code{name} is missing, \code{pathwaysResults} returns a \code{list} in which each item contains the results of a single pathway analysis.
If \code{name} is given, \code{pathwaysResults} returns a \code{\linkS4class{DataFrame}} that contains the results of a single pathway analysis.
}
\description{
\code{pathwaysResults} returns either all pathway analysis results stored in \code{object}
or a single pathway analysis result by name.

\code{pathwaysResultsNames} returns the names of pathway analysis results embedded in \code{object}.
}
\examples{
#' library("iSEEpathways")
library("org.Hs.eg.db")
library("fgsea")
library("iSEE")

# Example data ----

## Pathways
pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)
len_pathways <- lengths(pathways)
pathways <- pathways[len_pathways > 15 & len_pathways < 500]

## Features
set.seed(1)
# simulate a score for all genes found across all pathways
feature_stats <- rnorm(length(unique(unlist(pathways))))
names(feature_stats) <- unique(unlist(pathways))
# arbitrarily select a pathway to simulate enrichment
pathway_id <- "GO:0046324"
pathway_genes <- pathways[[pathway_id]]
# increase score of genes in the selected pathway to simulate enrichment
feature_stats[pathway_genes] <- feature_stats[pathway_genes] + 1

# fgsea ----

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
  stats    = feature_stats,
  minSize  = 15,
  maxSize  = 500)
head(fgseaRes[order(pval), ])

# iSEE ---

ngenes <- length(feature_stats)
cnts <- matrix(rnbinom(n=ngenes*2, mu=100, size=1/0.5), nrow=ngenes)
rownames(cnts) <- names(feature_stats)
se <- SummarizedExperiment(assay = list(counts = cnts))

se <- embedPathwaysResults(fgseaRes, se, name = "fgsea", class = "fgsea", pathwayType = "GO",
  pathwaysList = pathways, featuresStats = feature_stats)

##
# List result names ---
##

pathwaysResultsNames(se)

##
# Extract results ---
##

pathwaysResults(se)
pathwaysResults(se, "fgsea")
}
