% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwaysTable-class.R
\docType{methods}
\name{PathwaysTable-class}
\alias{PathwaysTable-class}
\alias{PathwaysTable}
\title{The PathwaysTable class}
\description{
The \code{PathwaysTable} is a \linkS4class{Panel} where each row represents a set of features (i.e., rows).
Selections in this panel can be transmitted to other row-oriented panels.
}
\examples{
x <- PathwaysTable(ResultName="fgsea")
}
