% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-simulate.R
\name{simulateExampleData}
\alias{simulateExampleData}
\title{Simulate Example Data}
\usage{
simulateExampleData(
  n_pathways = 5000,
  n_features = 15000,
  pathway_sizes = 16:500,
  seed = 1
)
}
\arguments{
\item{n_pathways}{integer scalar, number of pathways to simulate.}

\item{n_features}{integer scalar, number of features to simulate.}

\item{pathway_sizes}{integer vector, possible sizes of pathway to sample from.}

\item{seed}{a single value, passed to \code{\link[=set.seed]{set.seed()}}.}
}
\value{
A list of two elements:
\describe{
\item{pathwaysList}{A named list of dummy pathways.
Names represent pathway identifiers;
values represent character vectors of feature identifiers.}
}
\item{featuresStat}{A named numeric vector.
Names represent feature identifiers;
values represent 'scores' (e.g., log2 fold-change)}
}
\description{
Simulate Example Data
}
\examples{
simulated_data <- simulateExampleData()
head(lengths(simulated_data$pathwaysList))
head(simulated_data$featuresStat)
}
