% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/utils-SummarizedExperiment.R
\docType{methods}
\name{utils-SummarizedExperiment}
\alias{utils-SummarizedExperiment}
\alias{embedPathwaysResults}
\alias{embedPathwaysResults,ANY-method}
\alias{embedPathwaysResultsMethods}
\alias{embedPathwaysResults,data.frame-method}
\title{Generics for Embbedding Pathway Analysis Results into a SummarizedExperiment Object}
\format{
An object of class \code{character} of length 1.
}
\usage{
embedPathwaysResults(x, se, name, ...)

embedPathwaysResultsMethods

\S4method{embedPathwaysResults}{data.frame}(x, se, name, class, ...)
}
\arguments{
\item{x}{Object to be embedded.}

\item{se}{A \linkS4class{SummarizedExperiment} object.}

\item{name}{Identifier for the embedded object.}

\item{...}{Arguments passed to and from other methods.}

\item{class}{Class to use for embedding \code{x}. Only used when \code{class(x)} does
not uniquely identify the package that generated the object.}
}
\value{
An updated \linkS4class{SummarizedExperiment} object that contains the
embedded object.
}
\description{
An overview of the generics for embedding pathway analysis results into a \linkS4class{SummarizedExperiment} object, in a format compatible with \pkg{iSEEpathways}.
}
\section{Definitions}{

\itemize{
\item \code{embedPathwaysResults(x, se, name, ...)} embeds the results \code{x} in the \linkS4class{SummarizedExperiment} \code{se}.
}
}

\examples{
embedPathwaysResultsMethods

showMethods(embedPathwaysResults)
}
\author{
Kevin Rue-Albrecht
}
\keyword{datasets}
