% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FgseaEnrichmentPlot-class.R
\docType{methods}
\name{FgseaEnrichmentPlot-class}
\alias{FgseaEnrichmentPlot-class}
\alias{FgseaEnrichmentPlot}
\alias{initialize,FgseaEnrichmentPlot-method}
\alias{.cacheCommonInfo,FgseaEnrichmentPlot-method}
\alias{.createObservers,FgseaEnrichmentPlot-method}
\alias{.defineDataInterface,FgseaEnrichmentPlot-method}
\alias{.fullName,FgseaEnrichmentPlot-method}
\alias{.generateTable,FgseaEnrichmentPlot-method}
\alias{.multiSelectionActive,FgseaEnrichmentPlot-method}
\alias{.multiSelectionCommands,FgseaEnrichmentPlot-method}
\alias{.multiSelectionDimension,FgseaEnrichmentPlot-method}
\alias{.isBrushable,FgseaEnrichmentPlot-method}
\alias{.panelColor,FgseaEnrichmentPlot-method}
\alias{.refineParameters,FgseaEnrichmentPlot-method}
\alias{.showSelectionDetails,FgseaEnrichmentPlot-method}
\title{The FgseaEnrichmentPlot class}
\description{
The \code{FgseaEnrichmentPlot} is a \linkS4class{Panel} where each row represents a set of features (i.e., rows).
Selections in this panel can be transmitted to other row-oriented panels.
}
\examples{
x <- FgseaEnrichmentPlot(ResultName="fgsea", PathwayId="GO:0000002")
}
