% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfgseaResults-class.R
\docType{methods}
\name{iSEEfgseaResults-class}
\alias{iSEEfgseaResults-class}
\alias{iSEEfgseaResults}
\alias{embedPathwaysResults,iSEEfgseaResults-method}
\title{The iSEEfgseaResults class}
\description{
The \code{iSEEfgseaResults} class is used to provide an common interface to pathway analysis results produced by the \pkg{fgsea} package.
It provides methods to access the set of features in each pathway.
}
\details{
This class inherits all its slots directly from its parent class \linkS4class{DataFrame}.
}
\section{Constructor}{

\code{iSEEfgseaResults(data, row.names = rownames(data))} creates an instance of a \code{iSEEfgseaResults} class, with:

\describe{
\item{\code{data}}{A \code{data.frame} produced by \code{fgsea::fgsea()}.}
}
}

\section{Supported methods}{

\itemize{
\item \code{embedResults(x, se, name, class = "fgsea", ...)} embeds \code{x} in the column \code{name} of \code{metadata(se)[["iSEEpathways"]]}.
}
}

\examples{
library("org.Hs.eg.db")
library("fgsea")
library("SummarizedExperiment")

##
# Prepare pathways
##

pathways <- select(org.Hs.eg.db, keys(org.Hs.eg.db, "SYMBOL"), c("GOALL"), keytype = "SYMBOL")
pathways <- subset(pathways, ONTOLOGYALL == "BP")
pathways <- unique(pathways[, c("SYMBOL", "GOALL")])
pathways <- split(pathways$SYMBOL, pathways$GOALL)

##
# Prepare gene statistics
##
gene_symbols <- unique(unlist(pathways))
gene_stats <- rnorm(length(gene_symbols))
names(gene_stats) <- gene_symbols

##
# Run pathway analysis ----
##

set.seed(42)
fgseaRes <- fgsea(pathways = pathways,
                  stats    = gene_stats,
                  minSize  = 15,
                  maxSize  = 500)
head(fgseaRes[order(pval), ])

##
# iSEEfgseaResults ----
##

# Simulate the original SummarizedExperiment object
se <- SummarizedExperiment()

# Package the results in a iSEEfgseaResults object
iseefgsea_table <- iSEEfgseaResults(fgseaRes)

se <- embedPathwaysResults(iseefgsea_table, se, name = "fgsea")

se

##
# Methods ----
##

## TODO
}
\author{
Kevin Rue-Albrecht
}
