% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEfgseaResults-class.R
\name{iSEEpathwaysResults-class}
\alias{iSEEpathwaysResults-class}
\alias{show,iSEEpathwaysResults-method}
\alias{pathwayType,iSEEpathwaysResults-method}
\title{The iSEEpathwaysResults class}
\description{
The iSEEpathwaysResults class represents an undefined resource.
}
\section{Slot overview}{

\itemize{
\item \code{pathwayType}, a character scalar specifying the type of pathway (e.g., \code{"GO"}).
}
}

\section{Supported methods}{

In the following code snippets, \code{x} is an instance of a \code{\linkS4class{iSEEpathwaysResults}} class.
Refer to the documentation for each method for more details on the remaining arguments.

\itemize{
\item \code{\link{pathwayType}(x)} returns the type of pathways stored in \code{x}.
}
}

\examples{
new("iSEEpathwaysResults", pathwayType = "GO")
}
\author{
Kevin Rue-Albrecht
}
