% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{pathway-generics}
\alias{pathway-generics}
\alias{pathwayType}
\title{Generics for Pathway Analysis Results}
\description{
An overview of the generics for accessing common pieces of information in pathway analysis results.
}
\section{Definitions}{

\itemize{
\item \code{pathwayType(x)} returns a character scalar indicating the type of pathways analysed.
}
}

\examples{
showMethods(pathwayType)
}
\author{
Kevin Rue-Albrecht
}
