% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwaysTable-class.R
\docType{methods}
\name{PathwaysTable-class}
\alias{PathwaysTable-class}
\alias{PathwaysTable}
\alias{initialize,PathwaysTable-method}
\alias{.cacheCommonInfo,PathwaysTable-method}
\alias{.createObservers,PathwaysTable-method}
\alias{.defineDataInterface,PathwaysTable-method}
\alias{.defineInterface,PathwaysTable-method}
\alias{.fullName,PathwaysTable-method}
\alias{.generateTable,PathwaysTable-method}
\alias{.multiSelectionActive,PathwaysTable-method}
\alias{.multiSelectionCommands,PathwaysTable-method}
\alias{.multiSelectionDimension,PathwaysTable-method}
\alias{.panelColor,PathwaysTable-method}
\alias{.refineParameters,PathwaysTable-method}
\alias{.showSelectionDetails,PathwaysTable-method}
\title{The PathwaysTable class}
\description{
The \code{PathwaysTable} is a \linkS4class{Panel} where each row represents a set of features (i.e., rows).
Selections in this panel can be transmitted to other row-oriented panels.
}
\examples{
PathwaysTable(ResultName="fgsea")
}
